+++ src/components/recorder/UseMediaRecorder.jsx
@@ -260,7 +260,10 @@ function tryValidateMediaRecorderConstraints(constraints) {
  * @param {Array<string>} possibleTypes
  */
 export function getMediaRecorderSupportedMimeType(possibleTypes) {
-  if (typeof window.MediaRecorder === 'undefined') {
+  if (
+    typeof window === 'undefined' ||
+    typeof window.MediaRecorder === 'undefined'
+  ) {
     return '';
   }
 
@@ -271,4 +274,22 @@ export function getMediaRecorderSupportedMimeType(possibleTypes) {
   }
 
   return '';
+}
+
+/**
+ * @param {string} mimeType
+ */
+export function getVideoFileExtensionByMIMEType(mimeType) {
+  switch (mimeType) {
+    case 'video/mp4':
+      return '.mp4';
+    case 'video/quicktime':
+      return '.mov';
+    case 'video/webm':
+      return '.webm';
+    case 'video/ogg':
+      return '.ogg';
+    default:
+      return '';
+  }
 }