+++ src/components/workspace/DocumentPanel.jsx
@@ -1,11 +1,14 @@
 import { useState } from 'react';
 
+import AddToDriveIcon from '@material-symbols/svg-400/rounded/add_to_drive.svg';
 import BackIcon from '@material-symbols/svg-400/rounded/arrow_back-fill.svg';
+import CloudOffIcon from '@material-symbols/svg-400/rounded/cloud_off.svg';
 import DeleteIcon from '@material-symbols/svg-400/rounded/delete.svg';
 import ExportIcon from '@material-symbols/svg-400/rounded/export_notes-fill.svg';
 import MenuIcon from '@material-symbols/svg-400/rounded/menu.svg';
 
 import FancyButton from '@/lib/FancyButton';
+import { useGAPILogin, useGAPILogout } from '@/lib/googleapi/GoogleAPIContext';
 import { useDocumentStore } from '@/stores/DocumentStoreContext';
 import {
   useCurrentRecorder,
@@ -13,6 +16,7 @@ import {
   useSetUserCursor,
 } from '@/stores/UserStoreContext';
 
+import ExportSignal from './ExportSignal';
 import SceneList from './SceneList';
 
 /**
@@ -23,6 +27,9 @@ export default function DocumentPanel({ documentId }) {
   const setUserCursor = useSetUserCursor();
   const deleteDocument = useDocumentStore((ctx) => ctx.deleteDocument);
   const [isMenuOpen, setMenuOpen] = useState(false);
+
+  const login = useGAPILogin();
+  const logout = useGAPILogout();
 
   function onMenu() {
     setMenuOpen((prev) => !prev);
@@ -36,12 +43,20 @@ export default function DocumentPanel({ documentId }) {
     }
   }
 
+  function onSync() {
+    login();
+  }
+
+  function onDesync() {
+    logout();
+  }
+
   return (
     <>
       <div className="fixed m-2 z-10 right-0 flex flex-row">
         <div
           className={
-            'flex flex-row' +
+            'absolute right-0 top-14 flex flex-col' +
             ' ' +
             (isMenuOpen ? 'opacity-100' : 'opacity-0 pointer-events-none')
           }>
@@ -54,7 +69,17 @@ export default function DocumentPanel({ documentId }) {
           <FancyButton className="bg-white ml-2" title="Export">
             <ExportIcon className="inline w-6 fill-current" />
           </FancyButton>
+          <FancyButton className="bg-white ml-2" title="Sync" onClick={onSync}>
+            <AddToDriveIcon className="inline w-6 fill-current" />
+          </FancyButton>
+          <FancyButton
+            className="bg-white ml-2"
+            title="Desync"
+            onClick={onDesync}>
+            <CloudOffIcon className="inline w-6 fill-current" />
+          </FancyButton>
         </div>
+        <ExportSignal />
         <FancyButton className="bg-white ml-2" onClick={onMenu}>
           <MenuIcon className="inline w-6 fill-current" />
         </FancyButton>