+++ src/lib/googleapi/Keys.js
@@ -1,0 +1,3 @@
+export const GAPI_API_KEY = 'AIzaSyCBjZLAwZmYFLqkdF8fJnrKZtI_D8jd9x4';
+export const GAPI_CLIENT_ID =
+  '697267305763-b4jlr8dbpvnfk6ljbc2rl3mabjqptl2u.apps.googleusercontent.com';